Refactor sandbox sync
We're now going via an intermediate "meta" bare git repo.
This is a) cleaner to work with on the host, since there are not all
kinds of cryptic sandbox-related remotes in the host repo. And b) it
will be easier to adapt this to sandboxes running on a remote host.